Cast size to uint32 in custom new operator

+++ src/leeds/base/memoryManager.cpp
@@ -56,12 +56,12 @@ namespace base
 
 void* operator new(size_t size)
 {
-	return base::cMainMemoryManager::Instance()->Allocate(size);
+	return base::cMainMemoryManager::Instance()->Allocate((uint32)size);
 }
 
 void* operator new[](size_t size)
 {
-	return base::cMainMemoryManager::Instance()->Allocate(size);
+	return base::cMainMemoryManager::Instance()->Allocate((uint32)size);
 }
 
 void operator delete(void* buf) noexcept
